requirements-cpu.txt
certifi
charset-normalizer==3.0.1
idna==3.4
numpy==1.24.2
Pillow==9.4.0
PyYAML==6.0
requests==2.28.2
scipy==1.10.1
torch==1.13.1
torchaudio==0.13.1
torchvision==0.14.1
typing_extensions==4.5.0
urllib3==1.26.14
wincertstore==0.2
